requirements.txt
streamlit
faiss-cpu
sentence-transformers
PyPDF2
openai
fastapi
uvicorn
python-multipart
